[patch] refactor divisions context
