+++ lib/ehealth/declarations/api.ex
@@ -13,8 +13,8 @@ defmodule EHealth.Declarations.API do
   alias EHealth.PRM.LegalEntities.Schema, as: LegalEntity
   alias EHealth.PRM.Employees
   alias EHealth.PRM.Employees.Schema, as: Employee
-  alias EHealth.PRM.Divisions
-  alias EHealth.PRM.Divisions.Schema, as: Division
+  alias EHealth.Divisions
+  alias EHealth.Divisions.Division
 
   def get_declarations(params, headers) do
     with {:ok, resp} <- OPS.get_declarations(params, headers),
@@ -23,7 +23,7 @@ defmodule EHealth.Declarations.API do
          employee_ids <- list_to_param(related_ids["employee_ids"]),
          legal_entity_ids <- list_to_param(related_ids["legal_entity_ids"]),
          person_ids <- list_to_param(related_ids["person_ids"]),
-         %Page{} = divisions <- Divisions.get_divisions(%{ids: division_ids}),
+         %Page{} = divisions <- Divisions.list(%{ids: division_ids}),
          %Page{} = employees <- Employees.get_employees(%{ids: employee_ids}),
          %Page{} = legal_entities <- LegalEntities.get_legal_entities(%{ids: legal_entity_ids}),
          {:ok, persons} <- preload_persons(person_ids, headers),
@@ -124,7 +124,7 @@ defmodule EHealth.Declarations.API do
     with :ok          <- check_declaration_access(legal_entity_id, headers),
          person       <- load_relation(MPI, :person, declaration["person_id"], headers),
          legal_entity <- LegalEntities.get_legal_entity_by_id(legal_entity_id),
-         division     <- Divisions.get_division_by_id(declaration["division_id"]),
+         division     <- Divisions.get_by_id(declaration["division_id"]),
          employee     <- Employees.get_employee_by_id(declaration["employee_id"]),
          declaration  <- merge_related_data(declaration, person, legal_entity, division, employee),
          response     <- render_declaration(declaration),
