+++ lib/ehealth/declarations/view.ex
@@ -5,7 +5,7 @@ defmodule EHealth.Declarations.View do
   alias EHealth.PRM.LegalEntities.Schema, as: LegalEntity
   alias EHealth.PRM.Employees.Schema, as: Employee
   alias EHealth.PRM.Parties.Schema, as: Party
-  alias EHealth.PRM.Divisions.Schema, as: Division
+  alias EHealth.Divisions.Division
 
   def render_declarations(declarations) do
     Enum.map(declarations, &render_declaration(&1, :list))
