+++ lib/ehealth/divisions/uaddress.ex
@@ -3,7 +3,7 @@ defmodule EHealth.Divisions.UAddress do
   Service layer for synchronous update of the UAddress settlement and divisions
   """
   alias EHealth.API.UAddress
-  alias EHealth.PRM.Divisions
+  alias EHealth.Divisions
 
   import EHealth.Utils.Connection, only: [get_consumer_id: 1]
 
@@ -51,7 +51,7 @@ defmodule EHealth.Divisions.UAddress do
 
   defp api_update_divisions(%{"id" => id, "mountain_group" => group} = data, headers, settlement) do
     consumer_id = get_consumer_id(headers)
-    result = Divisions.update_divisions_mountain_group(%{settlement_id: id, mountain_group: group}, consumer_id)
+    result = Divisions.update_mountain_group(%{settlement_id: id, mountain_group: group}, consumer_id)
     case result do
       %Ecto.Changeset{valid?: false} -> result
       {:error, _failed_operation, failed_value, _changes_so_far} -> failed_value
